Move state functions to a util
